+++ A4/PhongMaterial.cpp
@@ -1,9 +1,9 @@
 #include "PhongMaterial.hpp"
 
-PhongMaterial::PhongMaterial(const glm::vec3& kd, const glm::vec3& ks, double shininess)
+PhongMaterial::PhongMaterial(const glm::dvec3& kd, const glm::dvec3& ks, double shininess)
 	: m_kd(kd)
   , m_ks(ks)
-  , m_refr(0.0f)
+  , m_refr(0.0)
 	, m_indexOfRefraction(1)
 	, m_shininess(shininess)
 {}
